requirements.txt
pandas
numpy
flask
python-dotenv
mysql_connector-python
pymysql
scikit-learn

#-e .